+++ ip/ip_test.go
@@ -1,11 +1,11 @@
-package server
+package ip
 
 import (
 	"net/http"
 	"testing"
 )
 
-func TestRemoteIPAddress(t *testing.T) {
+func TestResolve(t *testing.T) {
 	var tests = []struct {
 		header string
 		value  string
@@ -23,31 +23,11 @@ func TestRemoteIPAddress(t *testing.T) {
 	req := &http.Request{RemoteAddr: "192.168.0.1:55999", Header: header}
 	for _, test := range tests {
 		req.Header.Set(test.header, test.value)
-		got := remoteIPAddress(req)
+		got := Resolve(req)
 		if got != test.want {
 			t.Errorf("incorrect value, got: %s, want: %s", got, test.want)
 		}
 		req.Header.Set(test.header, "")
-	}
-}
-
-func TestHeaderValues(t *testing.T) {
-	expected := "192.168.0.1"
-
-	var tests = []struct {
-		input string
-		want  string
-	}{
-		{input: "192.168.0.1, 192.168.0.2", want: expected},
-		{input: "192.168.0.1,192.168.0.2", want: expected},
-		{input: "192.168.0.1", want: expected},
-	}
-
-	for _, test := range tests {
-		got := headerValues(test.input)
-		if got[0] != test.want {
-			t.Errorf("incorrect value, got: %s, want: %s", got[0], test.want)
-		}
 	}
 }