+++ server/handlers.go
@@ -2,6 +2,9 @@ package server
 
 import (
 	"net/http"
+	"strings"
+
+	"github.com/RedeployAB/gpip/ip"
 )
 
 // getIP handles incoming requests. With no provided headers, Accept: */*,
@@ -16,22 +19,20 @@ func (s *Server) getIP() http.Handler {
 			return
 		}
 		// Check accept header.
-		contentType := headerValues(r.Header.Get("accept"))[0]
-
+		contentType := strings.Split(strings.Replace(r.Header.Get("accept"), " ", "", -1), ",")[0]
 		var response func(w http.ResponseWriter, s string)
-		if len(contentType) == 0 || contentType == "application/json" || contentType == "*/*" {
+		switch contentType {
+		case "application/json", "*/*", "":
 			response = jsonResponse
-		} else if contentType == "text/plain" {
+		case "text/plain":
 			response = textResponse
-		} else {
+		default:
 			w.WriteHeader(http.StatusUnsupportedMediaType)
 			return
 		}
 
-		ip := remoteIPAddress(r)
-
 		w.Header().Set("Content-Type", contentType+"; charset=UTF-8")
 		w.WriteHeader(http.StatusOK)
-		response(w, ip)
+		response(w, ip.Resolve(r))
 	})
 }